feat: Added a new component - a stub - Refactored module with utils functions - Temporarily removed render info-trip-view
