+++ src/main.js
@@ -1,10 +1,8 @@
-import { render, RenderPosition } from './framework/render.js';
+import { render } from './framework/render.js';
 import PointsModel from './model/points-model.js';
 import EventsPresenter from './presenter/events-presenter.js';
 import FilterView from './view/filter-view.js';
-import InfoTripView from './view/info-trip-view.js';
 
-const infoTripElement = document.querySelector('.trip-main');
 const filterElement = document.querySelector('.trip-controls__filters');
 const eventsElement = document.querySelector('.trip-events');
 
@@ -14,7 +12,6 @@ const eventsPresenter = new EventsPresenter({
   pointsModel
 });
 
-render(new InfoTripView(), infoTripElement, RenderPosition.AFTERBEGIN);
 render(new FilterView(), filterElement);
 
 eventsPresenter.init();
