+++ src/utils/common.js
@@ -1,0 +1,5 @@
+function getRandomArrayElement(items) {
+  return items[Math.floor(Math.random() * items.length)];
+}
+
+export { getRandomArrayElement };
