+++ src/view/empty-points-view.js
@@ -1,0 +1,14 @@
+import AbstractView from '../framework/view/abstract-view.js';
+
+function createEmptyPointsTemplate() {
+  return `
+  <p class="trip-events__msg">
+    Click New Event to create your first point
+  </p>`;
+}
+
+export default class EmptyPointsView extends AbstractView {
+  get template() {
+    return createEmptyPointsTemplate();
+  }
+}
